OPT.BUG: calc_beamenvelope in transport lines
The optics.calc_beamenvelope function was always turning on
the cavity and the radiation, even if a fixed point is passed.
In this way, an exception occurs when propagating the envelope
in a transport line, because its cavity is always False and a
immutable property.
I changed the method to only turn on the cavity and rad if the
fixed point is not passed.