OPT.ENV.MNT: implement appropriate fix for envelope propagation on transfer lines.
The last fix was overloading the meaning of the fixed_point input, using it to decide whether the cavity and radiation states should be changed within the code.
With this fix these properties will always be changed, with exception of the cavity_on for transfer lines which do not have the harmonic_number set.

However, we could rethink this overall behavior of pyaccel, where functions change locally the radiation_on and cavity_on properties of the accelerator to make them compliant with what is expected.
Maybe, a better approach would be to raise errors and warnings of not compliant states and let the user change the properties accordingly.